+++ ui/src/routes/download/index.svelte
@@ -1,5 +1,5 @@
 <script context="module">
-	import {basename} from 'app/utils';
+	import {basename} from 'app/utils/assets';
 
 	export async function preload() {
 		return this.redirect(302, `/data/${basename}.zip`);